added pushMatrix() and popMatrix() to the render
