+++ java/src/com/C21751999/visual4/Visual4.java
@@ -1,7 +1,7 @@
 package com.C21751999.visual4;
 
 import com.C21460524.drawObjects.ExampleDraw;
-import com.C21751999.drawObjects.Flower;
+import com.C21751999.drawObjects.HanaSaku;
 
 import ddf.minim.AudioBuffer;
 import ddf.minim.AudioPlayer;
@@ -19,7 +19,7 @@ public class Visual4 extends VisualAbstractClass {
     //  Render Objects
     // Place Objects Here
     ExampleDraw exampleDraw;
-    Flower flo;
+    HanaSaku hana;
     //
     //
 
@@ -37,7 +37,7 @@ public class Visual4 extends VisualAbstractClass {
     // Draw Function for Visual 3
     public void drawVisual() {
         // Call Draw Functions in here
-        flo.render();
+        hana.render();
     } // End drawVisual2
 
 
@@ -45,6 +45,6 @@ public class Visual4 extends VisualAbstractClass {
     private void loadRenderObjects() {
         // Create your draw Objects here
         // exampleDraw = new ExampleDraw(this.pApplet, this.audioBuffer, this.windowWidth, this.windowHeight);
-        flo = new Flower(pApplet, audioBuffer, fft, windowWidth, windowHeight);
+        hana = new HanaSaku(pApplet, audioBuffer, fft, windowWidth, windowHeight);
     } // End void oadRenderObjects
 } // End clas Visual4
