+++ java/src/ie/tudublin/MusicVisualizer.java
@@ -87,7 +87,7 @@ public class MusicVisualizer extends PApplet {
         this.visual4 = new Visual4(this, this.audioBuffer, this.audioPlayer, this.windowWidth, this.windowHeight);
 
         visualList.add(visual4);
-        visualList.add(visual4);
+        visualList.add(visual2);
         visualList.add(visual2);
         visualList.add(visual1);
         visualList.add(visual1);
